Fix issue import users

+++ hserv/entity/dbSysUsers.php
@@ -757,7 +757,7 @@ class DbSysUsers extends DbEntityBase
                     }    
                 }
                 
-                if(count($remove)>0){
+                if(is_array($remove) && count($remove)>0){
                     $query = 'DELETE FROM sysUsrGrpLinks WHERE ugl_GroupID='.$groupID.' AND ugl_UserID in ('
                             .implode(',',$remove).')';
                     $res = $mysqli->query($query);
